Trait wasmer_vbus::VirtualBusProcess

pub trait VirtualBusProcess: VirtualBusScope + VirtualBusInvokable + Debug + Send + Sync + 'static {
    fn exit_code(&self) -> Option<u32>;

    fn stdin_fd(&self) -> Option<FileDescriptor>;

    fn stdout_fd(&self) -> Option<FileDescriptor>;

    fn stderr_fd(&self) -> Option<FileDescriptor>;
}

Required methods

fn exit_code(&self) -> Option<u32>

Returns the exit code if the instance has finished

fn stdin_fd(&self) -> Option<FileDescriptor>

Returns a file descriptor used to read the STDIN

fn stdout_fd(&self) -> Option<FileDescriptor>

Returns a file descriptor used to write to STDOUT

fn stderr_fd(&self) -> Option<FileDescriptor>

Returns a file descriptor used to write to STDERR